pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reassist"
version = "3.1.0"
description = "ReAssist — Research Intelligence Engine with Multi-Agent Pipeline"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    # Core
    "openai>=1.35.0",
    "python-dotenv>=1.0.0",

    # API integrations
    "arxiv>=2.0.0",
    "requests>=2.31.0",

    # PDF parsing
    "pdfplumber>=0.10.0",

    # Utilities
    "tiktoken>=0.5.0",

    # Web Server
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "pydantic>=2.4.2",

    # RAG / Vector
    "chromadb>=0.4.15",
    "langgraph>=0.5.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-openai>=0.3.0",

    # Database
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",

    # Auth
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",

    # Adaptive-Rag Extensions
    "tavily-python>=0.3.0",
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "jsonschema>=4.19.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
postgres = [
    "psycopg2-binary>=2.9.9",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
line-length = 120
target-version = "py310"
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
